Commit 06 mars 2024

+++ PetCare-Tracker/src/main/java/org/petcare/tracker/project/Implementations/AnimalImplementation.java
@@ -1,9 +1,11 @@
 package org.petcare.tracker.project.Implementations;
 
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.petcare.tracker.project.Models.Animal;
 import org.petcare.tracker.project.Models.Owner;
 import org.petcare.tracker.project.Repositories.AnimalRepository;
+import org.petcare.tracker.project.Repositories.AppointmentRepository;
 import org.petcare.tracker.project.Repositories.OwnerRepository;
 import org.petcare.tracker.project.Services.AnimalService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
 public class AnimalImplementation implements AnimalService {
@@ -21,6 +24,11 @@ public class AnimalImplementation implements AnimalService {
 
     @Autowired
     private OwnerRepository ownerRepository;
+
+    @Autowired
+    private AppointmentRepository appointmentRepository;
+
+
 
     // Implementation for getting animal with owner id
     @Override
@@ -46,11 +54,13 @@ public class AnimalImplementation implements AnimalService {
     }
 
     @Override
+    @Transactional
     public Animal updateAnimal(Animal updatedAnimal) {
         Animal existingAnimal = animalRepository.findById(updatedAnimal.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Animal not found with ID: " + updatedAnimal.getId()));
 
         // Update fields of the existing animal
+        existingAnimal.setOwners(existingAnimal.getOwners());
         existingAnimal.setName(updatedAnimal.getName());
         existingAnimal.setRace(updatedAnimal.getRace());
         existingAnimal.setGender(updatedAnimal.getGender());
@@ -62,13 +72,30 @@ public class AnimalImplementation implements AnimalService {
         existingAnimal.setNotes(updatedAnimal.getNotes());
         existingAnimal.setPicture(updatedAnimal.getPicture());
 
-        // Save the updated doctor back to the database
-        return animalRepository.save(updatedAnimal);
+        // Save the updated animal back to the database
+        return animalRepository.save(existingAnimal);
     }
 
     @Override
+    @Transactional
     public void deleteAnimal(Long animalId) {
-        animalRepository.deleteById(animalId);
+        Optional<Animal> animalOptional = animalRepository.findById(animalId);
+        if (animalOptional.isPresent()) {
+            Animal animal = animalOptional.get();
+
+            // Delete associated appointments
+            appointmentRepository.deleteByAnimalId(animalId);
+
+            // Dissociate from owners
+            for (Owner owner : animal.getOwners()) {
+                owner.getAnimals().remove(animal);
+                ownerRepository.save(owner);
+            }
+
+            animalRepository.deleteById(animalId);
+        } else {
+            throw new EntityNotFoundException("Animal not found with id: " + animalId);
+        }
     }
 
 }
